+++ src/tunnel/detail/platform_linux/layer_linux.hpp
@@ -13,13 +13,16 @@
 #include <sys/types.h>
 #include <system_error>
 
-#include "../detail/log.hpp"
+#include "../../log_level.hpp"
+
+#include "../log.hpp"
+#include "../log_kind.hpp"
+
 #include "scoped_file_descriptor.hpp"
 
-#include "../log_kind.hpp"
-#include "../log_level.hpp"
-
 namespace tunnel
+{
+namespace detail
 {
 namespace platform_linux
 {
@@ -39,18 +42,16 @@ struct layer_linux : public Super
         if (fd < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::open,
-                tunnel::detail::log::str{"path", path.c_str()},
-                tunnel::detail::log::integer{"flags", flags},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::open,
+                          log::str{"path", path.c_str()},
+                          log::integer{"flags", flags},
+                          log::str{"error", error.message().c_str()});
             return scoped_file_descriptor{};
         }
 
         Super::do_log(log_level::debug, log_kind::open,
-                      tunnel::detail::log::str{"path", path.c_str()},
-                      tunnel::detail::log::integer{"flags", flags},
-                      tunnel::detail::log::integer{"fd", fd});
+                      log::str{"path", path.c_str()},
+                      log::integer{"flags", flags}, log::integer{"fd", fd});
 
         return scoped_file_descriptor{fd};
     }
@@ -65,21 +66,19 @@ struct layer_linux : public Super
         if (fd < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::socket,
-                tunnel::detail::log::integer{"domain", domain},
-                tunnel::detail::log::integer{"type", type},
-                tunnel::detail::log::integer{"protocol", protocol},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::socket,
+                          log::integer{"domain", domain},
+                          log::integer{"type", type},
+                          log::integer{"protocol", protocol},
+                          log::str{"error", error.message().c_str()});
 
             return scoped_file_descriptor{};
         }
 
-        Super::do_log(log_level::debug, log_kind::socket,
-                      tunnel::detail::log::integer{"domain", domain},
-                      tunnel::detail::log::integer{"type", type},
-                      tunnel::detail::log::integer{"protocol", protocol},
-                      tunnel::detail::log::integer{"fd", fd});
+        Super::do_log(
+            log_level::debug, log_kind::socket, log::integer{"domain", domain},
+            log::integer{"type", type}, log::integer{"protocol", protocol},
+            log::integer{"fd", fd});
 
         return scoped_file_descriptor{fd};
     }
@@ -95,15 +94,14 @@ struct layer_linux : public Super
         if (::ioctl(fd.native_handle(), request, arg) == -1)
         {
             error.assign(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::ioctl,
-                tunnel::detail::log::uinteger{"request", request},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::ioctl,
+                          log::uinteger{"request", request},
+                          log::str{"error", error.message().c_str()});
         }
 
         Super::do_log(log_level::debug, log_kind::ioctl,
-                      tunnel::detail::log::uinteger{"request", request},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()});
+                      log::uinteger{"request", request},
+                      log::integer{"fd", fd.native_handle()});
     }
 
     auto send(const scoped_file_descriptor& fd, void* data, uint32_t size,
@@ -119,19 +117,17 @@ struct layer_linux : public Super
         if (res < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::send,
-                tunnel::detail::log::uinteger{"size", size},
-                tunnel::detail::log::integer{"flags", flags},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::send,
+                          log::uinteger{"size", size},
+                          log::integer{"flags", flags},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
         Super::do_log(log_level::debug, log_kind::send,
-                      tunnel::detail::log::uinteger{"size", size},
-                      tunnel::detail::log::integer{"flags", flags},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()},
-                      tunnel::detail::log::integer{"res", res});
+                      log::uinteger{"size", size}, log::integer{"flags", flags},
+                      log::integer{"fd", fd.native_handle()},
+                      log::integer{"res", res});
 
         return res;
     }
@@ -149,19 +145,17 @@ struct layer_linux : public Super
         if (res < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::recv,
-                tunnel::detail::log::uinteger{"size", size},
-                tunnel::detail::log::integer{"flags", flags},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::recv,
+                          log::uinteger{"size", size},
+                          log::integer{"flags", flags},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
         Super::do_log(log_level::debug, log_kind::recv,
-                      tunnel::detail::log::uinteger{"size", size},
-                      tunnel::detail::log::integer{"flags", flags},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()},
-                      tunnel::detail::log::integer{"res", res});
+                      log::uinteger{"size", size}, log::integer{"flags", flags},
+                      log::integer{"fd", fd.native_handle()},
+                      log::integer{"res", res});
 
         return res;
     }
@@ -179,17 +173,15 @@ struct layer_linux : public Super
         if (res < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::bind,
-                tunnel::detail::log::uinteger{"addrlen", addrlen},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::bind,
+                          log::uinteger{"addrlen", addrlen},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
-        Super::do_log(log_level::debug, log_kind::bind,
-                      tunnel::detail::log::uinteger{"addrlen", addrlen},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()},
-                      tunnel::detail::log::integer{"res", res});
+        Super::do_log(
+            log_level::debug, log_kind::bind, log::uinteger{"addrlen", addrlen},
+            log::integer{"fd", fd.native_handle()}, log::integer{"res", res});
 
         return res;
     }
@@ -205,10 +197,9 @@ struct layer_linux : public Super
         if (current_offset < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::size,
-                tunnel::detail::log::integer{"current_offset", current_offset},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::size,
+                          log::integer{"current_offset", current_offset},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
@@ -217,10 +208,9 @@ struct layer_linux : public Super
         if (offset < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::size,
-                tunnel::detail::log::integer{"offset", offset},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::size,
+                          log::integer{"offset", offset},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
@@ -230,16 +220,15 @@ struct layer_linux : public Super
         if (set_offset < 0)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::size,
-                tunnel::detail::log::integer{"set_offset", set_offset},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::size,
+                          log::integer{"set_offset", set_offset},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
         Super::do_log(log_level::debug, log_kind::size,
-                      tunnel::detail::log::integer{"offset", offset},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()});
+                      log::integer{"offset", offset},
+                      log::integer{"fd", fd.native_handle()});
 
         return offset;
     }
@@ -257,20 +246,19 @@ struct layer_linux : public Super
         if (res == -1)
         {
             error = std::error_code(errno, std::generic_category());
-            Super::do_log(
-                log_level::error, log_kind::read,
-                tunnel::detail::log::uinteger{"size", size},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::read,
+                          log::uinteger{"size", size},
+                          log::str{"error", error.message().c_str()});
             return 0;
         }
 
-        Super::do_log(log_level::debug, log_kind::read,
-                      tunnel::detail::log::uinteger{"size", size},
-                      tunnel::detail::log::integer{"fd", fd.native_handle()},
-                      tunnel::detail::log::integer{"res", res});
+        Super::do_log(
+            log_level::debug, log_kind::read, log::uinteger{"size", size},
+            log::integer{"fd", fd.native_handle()}, log::integer{"res", res});
 
         return res;
     }
 };
+}
 }
 }