+++ src/tunnel/detail/platform_linux/layer_netdevice.hpp
@@ -14,14 +14,17 @@
 #include <sys/ioctl.h>
 #include <system_error>
 
-#include "../detail/log.hpp"
+#include "../log.hpp"
+#include "../log_kind.hpp"
+
 #include "error.hpp"
 #include "scoped_file_descriptor.hpp"
 
-#include "../log_kind.hpp"
-#include "../log_level.hpp"
+#include "../../log_level.hpp"
 
 namespace tunnel
+{
+namespace detail
 {
 namespace platform_linux
 {
@@ -53,11 +56,9 @@ public:
         if (interface_name.size() > IFNAMSIZ - 1)
         {
             error = make_error_code(linux_error::interface_name_too_long);
-            Super::do_log(
-                log_level::error, log_kind::interface_renamed,
-                tunnel::detail::log::str{"interface_name",
-                                         interface_name.c_str()},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::interface_renamed,
+                          log::str{"interface_name", interface_name.c_str()},
+                          log::str{"error", error.message().c_str()});
             return;
         }
 
@@ -94,9 +95,8 @@ public:
             return;
         }
 
-        Super::do_log(
-            log_level::state, log_kind::interface_renamed,
-            tunnel::detail::log::str{"interface_name", interface_name.c_str()});
+        Super::do_log(log_level::debug, log_kind::interface_renamed,
+                      log::str{"interface_name", interface_name.c_str()});
 
         if (iface_up)
         {
@@ -124,8 +124,8 @@ public:
 
         bool is_if_up = (ifr.ifr_flags & IFF_UP) != 0;
 
-        Super::do_log(log_level::state, log_kind::is_up,
-                      tunnel::detail::log::boolean{"is_up", is_if_up});
+        Super::do_log(log_level::debug, log_kind::is_up,
+                      log::boolean{"is_up", is_if_up});
 
         return is_if_up;
     }
@@ -151,13 +151,13 @@ public:
         ifr.ifr_flags |= IFF_UP;
         Super::ioctl(m_dev_fd, SIOCSIFFLAGS, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::interface_up);
+        Super::do_log(log_level::debug, log_kind::interface_up);
     }
 
     auto is_down(std::error_code& error) const -> bool
     {
         assert(!error);
-        Super::do_log(log_level::state, log_kind::is_down);
+        Super::do_log(log_level::debug, log_kind::is_down);
         return !is_up(error);
     }
 
@@ -182,7 +182,7 @@ public:
         ifr.ifr_flags &= ~IFF_UP;
         Super::ioctl(m_dev_fd, SIOCSIFFLAGS, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::interface_down);
+        Super::do_log(log_level::debug, log_kind::interface_down);
     }
 
     void set_mtu(uint32_t mtu, std::error_code& error) const
@@ -192,10 +192,9 @@ public:
         if (mtu < ETH_HLEN || mtu > 65535)
         {
             error = make_error_code(linux_error::mtu_too_large);
-            Super::do_log(
-                log_level::error, log_kind::set_mtu,
-                tunnel::detail::log::uinteger{"mtu", mtu},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::set_mtu,
+                          log::uinteger{"mtu", mtu},
+                          log::str{"error", error.message().c_str()});
             return;
         }
 
@@ -210,8 +209,8 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCSIFMTU, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::set_mtu,
-                      tunnel::detail::log::integer{"mtu", mtu});
+        Super::do_log(log_level::debug, log_kind::set_mtu,
+                      log::integer{"mtu", mtu});
     }
 
     auto mtu(std::error_code& error) const -> uint32_t
@@ -225,8 +224,8 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCGIFMTU, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::interface_mtu,
-                      tunnel::detail::log::integer{"mtu", ifr.ifr_mtu});
+        Super::do_log(log_level::debug, log_kind::interface_mtu,
+                      log::integer{"mtu", ifr.ifr_mtu});
 
         return ifr.ifr_mtu;
     }
@@ -266,7 +265,7 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCADDRT, &route, error);
 
-        Super::do_log(log_level::state, log_kind::enable_default_route);
+        Super::do_log(log_level::debug, log_kind::enable_default_route);
     }
 
     void disable_default_route(std::error_code& error) const
@@ -304,7 +303,7 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCDELRT, &route, error);
 
-        Super::do_log(log_level::state, log_kind::disable_default_route);
+        Super::do_log(log_level::debug, log_kind::disable_default_route);
     }
 
     auto ipv4(std::error_code& error) const -> std::string
@@ -327,9 +326,8 @@ public:
 
         struct sockaddr_in* addr_in = (struct sockaddr_in*)&ifr.ifr_addr;
 
-        Super::do_log(
-            log_level::state, log_kind::interface_ipv4,
-            tunnel::detail::log::str{"ip", ::inet_ntoa(addr_in->sin_addr)});
+        Super::do_log(log_level::debug, log_kind::interface_ipv4,
+                      log::str{"ip", ::inet_ntoa(addr_in->sin_addr)});
 
         return ::inet_ntoa(addr_in->sin_addr);
     }
@@ -354,9 +352,8 @@ public:
 
         struct sockaddr_in* addr_in = (struct sockaddr_in*)&ifr.ifr_addr;
 
-        Super::do_log(log_level::state, log_kind::interface_ipv4_netmask,
-                      tunnel::detail::log::str{"netmask",
-                                               ::inet_ntoa(addr_in->sin_addr)});
+        Super::do_log(log_level::debug, log_kind::interface_ipv4_netmask,
+                      log::str{"netmask", ::inet_ntoa(addr_in->sin_addr)});
 
         return ::inet_ntoa(addr_in->sin_addr);
     }
@@ -385,8 +382,8 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCSIFADDR, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::set_ipv4,
-                      tunnel::detail::log::str{"ip", address.c_str()});
+        Super::do_log(log_level::debug, log_kind::set_ipv4,
+                      log::str{"ip", address.c_str()});
     }
 
     void set_ipv4_netmask(const std::string& netmask,
@@ -412,8 +409,8 @@ public:
 
         Super::ioctl(m_dev_fd, SIOCSIFNETMASK, &ifr, error);
 
-        Super::do_log(log_level::state, log_kind::set_ipv4_netmask,
-                      tunnel::detail::log::str{"netmask", netmask.c_str()});
+        Super::do_log(log_level::debug, log_kind::set_ipv4_netmask,
+                      log::str{"netmask", netmask.c_str()});
     }
 
 private:
@@ -429,10 +426,9 @@ private:
         {
             error = std::error_code(errno, std::generic_category());
 
-            Super::do_log(
-                log_level::error, log_kind::make_sockaddr,
-                tunnel::detail::log::str{"ip", ip.c_str()},
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::make_sockaddr,
+                          log::str{"ip", ip.c_str()},
+                          log::str{"error", error.message().c_str()});
 
             return {};
         }
@@ -440,7 +436,7 @@ private:
         addr_in->sin_family = AF_INET;
 
         Super::do_log(log_level::debug, log_kind::make_sockaddr,
-                      detail::log::str{"ip", ip.c_str()});
+                      log::str{"ip", ip.c_str()});
 
         return addr;
     }
@@ -464,7 +460,10 @@ private:
         return ifr;
     }
 
-    private : scoped_file_descriptor m_dev_fd;
+    private :
+
+        scoped_file_descriptor m_dev_fd;
 };
+}
 }
 }