+++ src/tunnel/detail/platform_linux/layer_tun.hpp
@@ -29,12 +29,14 @@
 #include "error.hpp"
 #include "scoped_file_descriptor.hpp"
 
-#include "../detail/log.hpp"
-
+#include "../log.hpp"
 #include "../log_kind.hpp"
-#include "../log_level.hpp"
+
+#include "../../log_level.hpp"
 
 namespace tunnel
+{
+namespace detail
 {
 namespace platform_linux
 {
@@ -55,9 +57,8 @@ struct layer_tun : public Super
         {
             error = make_error_code(linux_error::interface_name_too_long);
 
-            Super::do_log(
-                log_level::error, log_kind::interface_created,
-                tunnel::detail::log::str{"error", error.message().c_str()});
+            Super::do_log(log_level::error, log_kind::interface_created,
+                          log::str{"error", error.message().c_str()});
 
             return;
         }
@@ -97,8 +98,8 @@ struct layer_tun : public Super
 
         Super::ioctl(m_tun_fd, TUNSETIFF, (void*)&ifr, error);
 
-        Super::do_log(log_level::state, log_kind::interface_created,
-                      tunnel::detail::log::str{"name", interface_name.c_str()});
+        Super::do_log(log_level::debug, log_kind::interface_created,
+                      log::str{"name", interface_name.c_str()});
     }
 
     auto owner(std::error_code& error) const -> std::string
@@ -107,8 +108,8 @@ struct layer_tun : public Super
 
         std::string o = read_property("owner", error);
 
-        Super::do_log(log_level::state, log_kind::owner,
-                      tunnel::detail::log::str{"owner", o.c_str()});
+        Super::do_log(log_level::debug, log_kind::owner,
+                      log::str{"owner", o.c_str()});
 
         if (error)
         {
@@ -139,8 +140,8 @@ struct layer_tun : public Super
         assert(!error);
         assert(!owner.empty());
 
-        Super::do_log(log_level::state, log_kind::set_owner,
-                      tunnel::detail::log::str{"owner", owner.c_str()});
+        Super::do_log(log_level::debug, log_kind::set_owner,
+                      log::str{"owner", owner.c_str()});
 
         struct passwd* pwd = getpwnam(owner.c_str());
 
@@ -158,8 +159,8 @@ struct layer_tun : public Super
     {
         std::string o = read_property("group", error);
 
-        Super::do_log(log_level::state, log_kind::group,
-                      tunnel::detail::log::str{"group", o.c_str()});
+        Super::do_log(log_level::debug, log_kind::group,
+                      log::str{"group", o.c_str()});
 
         if (error)
         {
@@ -191,8 +192,8 @@ struct layer_tun : public Super
         assert(!group.empty());
         assert(m_tun_fd);
 
-        Super::do_log(log_level::state, log_kind::set_group,
-                      tunnel::detail::log::str{"group", group.c_str()});
+        Super::do_log(log_level::debug, log_kind::set_group,
+                      log::str{"group", group.c_str()});
 
         struct group* grp = getgrnam(group.c_str());
 
@@ -218,8 +219,8 @@ struct layer_tun : public Super
 
         bool persistent = ifr.ifr_flags & IFF_PERSIST;
 
-        Super::do_log(log_level::state, log_kind::interface_is_persistent,
-                      tunnel::detail::log::boolean{"persistent", persistent});
+        Super::do_log(log_level::debug, log_kind::interface_is_persistent,
+                      log::boolean{"persistent", persistent});
         return persistent;
     }
 
@@ -228,8 +229,8 @@ struct layer_tun : public Super
         assert(m_tun_fd);
         assert(!error);
 
-        Super::do_log(log_level::state, log_kind::set_persistent,
-                      tunnel::detail::log::str{"", ""});
+        Super::do_log(log_level::debug, log_kind::set_persistent,
+                      log::str{"", ""});
 
         Super::ioctl(m_tun_fd, TUNSETPERSIST, (void*)1, error);
     }
@@ -239,8 +240,8 @@ struct layer_tun : public Super
         assert(m_tun_fd);
         assert(!error);
 
-        Super::do_log(log_level::state, log_kind::set_non_persistent,
-                      tunnel::detail::log::str{"", ""});
+        Super::do_log(log_level::debug, log_kind::set_non_persistent,
+                      log::str{"", ""});
 
         Super::ioctl(m_tun_fd, TUNSETPERSIST, (void*)0, error);
     }
@@ -255,8 +256,8 @@ struct layer_tun : public Super
         };
         Super::ioctl(m_tun_fd, TUNGETIFF, (void*)&ifr, error);
 
-        Super::do_log(log_level::state, log_kind::interface_name,
-                      tunnel::detail::log::str{"name", ifr.ifr_name});
+        Super::do_log(log_level::debug, log_kind::interface_name,
+                      log::str{"name", ifr.ifr_name});
 
         return ifr.ifr_name;
     }
@@ -265,9 +266,8 @@ struct layer_tun : public Super
     {
         assert(m_tun_fd);
 
-        Super::do_log(
-            log_level::state, log_kind::native_handle,
-            tunnel::detail::log::integer{"handle", m_tun_fd.native_handle()});
+        Super::do_log(log_level::debug, log_kind::native_handle,
+                      log::integer{"handle", m_tun_fd.native_handle()});
 
         return m_tun_fd.native_handle();
     }
@@ -318,5 +318,6 @@ private:
 private:
     scoped_file_descriptor m_tun_fd;
 };
+}
 }
 }