+++ test/src/detail/platform_linux/test_layer_linux.cpp
@@ -3,4 +3,4 @@
 //
 // Distributed under the "BSD License". See the accompanying LICENSE.rst file.
 
-#include <tunnel/platform_linux/layer_netlink_v4.hpp>
+#include <tunnel/detail/platform_linux/layer_linux.hpp>